requirements.txt
libsass==0.23.0
markdown==3.7.0
pillow==11.1.0
pystache==0.6.8
simple-http-server==0.22.3
pyyaml==6.0.2
